base webgl texture class
