+++ src/webgl/static-decal-texture-webgl.ts
@@ -1,40 +1,12 @@
-import { CanvasTexture, Color, LinearEncoding, RepeatWrapping, Texture } from 'three';
+import { Color } from 'three';
 import { Decal } from '../model/decal';
 import { getAssetUrl } from '../utils/network';
 import { BodyTexture } from '../3d/body/body-texture';
-import { MultiImageLoader, PromiseLoader } from '../utils/loader';
 import { Body } from '../model/body';
 import { PaintConfig } from '../model/paint-config';
 import { RocketConfig } from '../model/rocket-config';
-import { bindColor, bindEmptyTexture, createTextureFromImage, initShaderProgram, setRectangle } from '../utils/webgl';
-
-
-// language=GLSL
-const VERTEX_SHADER = `
-    attribute vec2 a_position;
-    attribute vec2 a_texCoord;
-
-    uniform vec2 u_resolution;
-
-    varying vec2 v_texCoord;
-
-    void main() {
-        // convert the rectangle from pixels to 0.0 to 1.0
-        vec2 zeroToOne = a_position / u_resolution;
-
-        // convert from 0->1 to 0->2
-        vec2 zeroToTwo = zeroToOne * 2.0;
-
-        // convert from 0->2 to -1->+1 (clipspace)
-        vec2 clipSpace = zeroToTwo - 1.0;
-
-        gl_Position = vec4(clipSpace * vec2(1, -1), 0, 1);
-
-        // pass the texCoord to the fragment shader
-        // The GPU will interpolate this value between points.
-        v_texCoord = a_texCoord;
-    }
-`;
+import { bindColor, bindEmptyTexture, createTextureFromImage } from '../utils/webgl';
+import { WebGLCanvasTexture } from './webgl-texture';
 
 // language=GLSL
 const FRAGMENT_SHADER = `
@@ -99,16 +71,13 @@ const FRAGMENT_SHADER = `
     }
 `;
 
-export class StaticDecalTextureWebGL implements BodyTexture {
+export class StaticDecalTextureWebGL extends WebGLCanvasTexture implements BodyTexture {
 
-  private readonly loader: PromiseLoader;
+  protected fragmentShader: string = FRAGMENT_SHADER;
 
-  private readonly baseUrl: string;
   private readonly rgbaMapUrl: string;
-  private readonly bodyBaseSkinUrl: string;
   private readonly bodyBlankSkinUrl: string;
 
-  private base: HTMLImageElement;
   private decalRgbaMap: HTMLImageElement;
   private bodyBlankSkin: HTMLImageElement;
 
@@ -117,29 +86,19 @@ export class StaticDecalTextureWebGL implements BodyTexture {
   private paint: Color;
   private bodyPaint: Color;
 
-  private canvas: OffscreenCanvas;
-  private gl: WebGLRenderingContext;
-  private program: WebGLProgram;
-
   private primaryLocation: WebGLUniformLocation;
   private bodyPaintLocation: WebGLUniformLocation;
   private accentLocation: WebGLUniformLocation;
   private decalPaintLocation: WebGLUniformLocation;
+  private rgbaMapLocation: WebGLUniformLocation;
+  private decalMapLocation: WebGLUniformLocation;
 
-  private baseTexture: WebGLTexture;
   private rgbaMapTexture: WebGLTexture;
   private decalMapTexture: WebGLTexture;
 
-  private texCoordBuffer: WebGLBuffer;
-  private positionBuffer: WebGLBuffer;
-
-  private texture: CanvasTexture;
-
   constructor(body: Body, decal: Decal, paints: PaintConfig, rocketConfig: RocketConfig) {
-    this.loader = new PromiseLoader(new MultiImageLoader(rocketConfig.textureFormat, rocketConfig.loadingManager));
-    this.baseUrl = getAssetUrl(decal.base_texture, rocketConfig);
+    super(decal.base_texture != undefined ? getAssetUrl(decal.base_texture, rocketConfig) : getAssetUrl(body.base_skin, rocketConfig), rocketConfig);
     this.rgbaMapUrl = getAssetUrl(decal.rgba_map, rocketConfig);
-    this.bodyBaseSkinUrl = getAssetUrl(body.base_skin, rocketConfig);
     this.bodyBlankSkinUrl = getAssetUrl(body.blank_skin, rocketConfig);
 
     this.primary = paints.primary;
@@ -149,16 +108,9 @@ export class StaticDecalTextureWebGL implements BodyTexture {
   }
 
   async load() {
-    const baseTask = this.loader.load(this.baseUrl != undefined ? this.baseUrl : this.bodyBaseSkinUrl);
+    const superTask = super.load();
     const rgbaMapTask = this.loader.load(this.rgbaMapUrl);
     const bodyBlankSkinTask = this.loader.load(this.bodyBlankSkinUrl);
-
-    const baseResult = await baseTask;
-
-    const width = baseResult.width;
-    const height = baseResult.height;
-
-    this.base = baseResult;
 
     const rgbaMapResult = await rgbaMapTask;
     if (rgbaMapResult != undefined) {
@@ -167,32 +119,13 @@ export class StaticDecalTextureWebGL implements BodyTexture {
 
     this.bodyBlankSkin = await bodyBlankSkinTask;
 
-    this.initWebGL(width, height);
-
-    // @ts-ignore
-    this.texture = new CanvasTexture(this.canvas);
-    this.texture.wrapS = RepeatWrapping;
-    this.texture.wrapT = RepeatWrapping;
-    this.texture.encoding = LinearEncoding;
-    this.texture.flipY = false;
-    this.update();
+    await superTask;
   }
 
-  private initWebGL(width: number, height: number) {
-    this.canvas = new OffscreenCanvas(width, height);
-
-    this.gl = this.canvas.getContext('webgl', {premultipliedAlpha: false});
-    this.program = initShaderProgram(this.gl, VERTEX_SHADER, FRAGMENT_SHADER);
-    this.gl.useProgram(this.program);
-
-    // look up where the vertex data needs to go.
-    const baseLocation = this.gl.getUniformLocation(this.program, 'u_base');
-    const rgbaMapLocation = this.gl.getUniformLocation(this.program, 'u_rgba_map');
-    const decalMapLocation = this.gl.getUniformLocation(this.program, 'u_decal_map');
-    const resolutionLocation = this.gl.getUniformLocation(this.program, 'u_resolution');
+  protected initWebGL(width: number, height: number) {
+    this.rgbaMapLocation = this.gl.getUniformLocation(this.program, 'u_rgba_map');
+    this.decalMapLocation = this.gl.getUniformLocation(this.program, 'u_decal_map');
     const isBlankLocation = this.gl.getUniformLocation(this.program, 'u_is_blank');
-    const positionLocation = this.gl.getAttribLocation(this.program, 'a_position');
-    const texCoordLocation = this.gl.getAttribLocation(this.program, 'a_texCoord');
 
     this.primaryLocation = this.gl.getUniformLocation(this.program, 'u_primary');
     this.bodyPaintLocation = this.gl.getUniformLocation(this.program, 'u_body_paint');
@@ -201,20 +134,11 @@ export class StaticDecalTextureWebGL implements BodyTexture {
 
     this.gl.uniform1i(isBlankLocation, this.decalRgbaMap != undefined ? 0 : 1);
 
-    // provide texture coordinates for the rectangle.
-    this.texCoordBuffer = this.gl.createBuffer();
-    this.gl.bindBuffer(this.gl.ARRAY_BUFFER, this.texCoordBuffer);
-    this.gl.bufferData(this.gl.ARRAY_BUFFER, new Float32Array([
-      0.0, 0.0,
-      1.0, 0.0,
-      0.0, 1.0,
-      0.0, 1.0,
-      1.0, 0.0,
-      1.0, 1.0]), this.gl.STATIC_DRAW);
-    this.gl.enableVertexAttribArray(texCoordLocation);
-    this.gl.vertexAttribPointer(texCoordLocation, 2, this.gl.FLOAT, false, 0, 0);
+    super.initWebGL(width, height);
+  }
 
-    this.baseTexture = createTextureFromImage(this.gl, this.base);
+  protected createTextures() {
+    super.createTextures();
     this.rgbaMapTexture = createTextureFromImage(this.gl, this.bodyBlankSkin);
 
     if (this.decalRgbaMap != undefined) {
@@ -222,29 +146,20 @@ export class StaticDecalTextureWebGL implements BodyTexture {
     } else {
       this.decalMapTexture = bindEmptyTexture(this.gl);
     }
+  }
 
-    this.gl.uniform1i(baseLocation, 0);
-    this.gl.uniform1i(rgbaMapLocation, 1);
-    this.gl.uniform1i(decalMapLocation, 2);
+  protected setTextureLocations() {
+    super.setTextureLocations();
+    this.gl.uniform1i(this.rgbaMapLocation, 1);
+    this.gl.uniform1i(this.decalMapLocation, 2);
+  }
 
-    this.gl.activeTexture(this.gl.TEXTURE0);
-    this.gl.bindTexture(this.gl.TEXTURE_2D, this.baseTexture);
+  protected bindTextures() {
+    super.bindTextures();
     this.gl.activeTexture(this.gl.TEXTURE1);
     this.gl.bindTexture(this.gl.TEXTURE_2D, this.rgbaMapTexture);
     this.gl.activeTexture(this.gl.TEXTURE2);
     this.gl.bindTexture(this.gl.TEXTURE_2D, this.decalMapTexture);
-
-    // set the resolution
-    this.gl.uniform2f(resolutionLocation, this.canvas.width, this.canvas.height);
-
-    // Create a buffer for the position of the rectangle corners.
-    this.positionBuffer = this.gl.createBuffer();
-    this.gl.bindBuffer(this.gl.ARRAY_BUFFER, this.positionBuffer);
-    this.gl.enableVertexAttribArray(positionLocation);
-    this.gl.vertexAttribPointer(positionLocation, 2, this.gl.FLOAT, false, 0, 0);
-
-    // Set a rectangle the same size as the image.
-    setRectangle(this.gl, 0, 0, this.base.width, this.base.height);
   }
 
   setBodyPaint(color: Color) {
@@ -267,37 +182,20 @@ export class StaticDecalTextureWebGL implements BodyTexture {
     this.update();
   }
 
-  private update() {
+  protected update() {
     bindColor(this.gl, this.primaryLocation, this.primary);
     bindColor(this.gl, this.accentLocation, this.accent);
     bindColor(this.gl, this.bodyPaintLocation, this.bodyPaint);
     bindColor(this.gl, this.decalPaintLocation, this.paint);
-
-    // Draw the rectangle.
-    this.gl.drawArrays(this.gl.TRIANGLES, 0, 6);
-    this.texture.needsUpdate = true;
+    super.update();
   }
 
   dispose() {
-    this.base = undefined;
+    super.dispose();
     this.decalRgbaMap = undefined;
     this.bodyBlankSkin = undefined;
 
-    this.gl.activeTexture(this.gl.TEXTURE0);
-    this.gl.bindTexture(this.gl.TEXTURE_2D, null);
-    this.gl.activeTexture(this.gl.TEXTURE1);
-    this.gl.bindTexture(this.gl.TEXTURE_2D, null);
-    this.gl.activeTexture(this.gl.TEXTURE2);
-    this.gl.bindTexture(this.gl.TEXTURE_2D, null);
-    this.gl.bindBuffer(this.gl.ARRAY_BUFFER, null);
-    this.gl.deleteTexture(this.baseTexture);
     this.gl.deleteTexture(this.rgbaMapTexture);
     this.gl.deleteTexture(this.decalMapTexture);
-    this.gl.deleteBuffer(this.texCoordBuffer);
-    this.gl.deleteBuffer(this.positionBuffer);
-  }
-
-  getTexture(): Texture {
-    return this.texture;
   }
 }
